Address review feedback & add attributes to ref
